requirements.txt
# PyTorch and related packages
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# PyTorch Lightning
pytorch-lightning>=2.0.0
lightning>=2.0.0

# Model libraries
timm>=0.9.0

# Data processing
pillow>=9.0.0
numpy>=1.24.0

# Logging and monitoring
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
